[TEST] Consolidate and expand GUI attachment tests
- Merge test_gui_attachments.py and test_gui_batch_attachments.py into a comprehensive suite.
- Add coverage for filename collision avoidance in batch extraction.
- Add coverage for edge cases: missing selection, invalid indices, and empty message text.
- Add coverage for I/O exception handling in save/extract methods.
- Achieve 100% test coverage for pyqwk/gui.py.
